Fix analysis

+++ projects/aprocess/test/is_process_running.cpp
@@ -3,7 +3,10 @@ import awinapi;
 
 [[nodiscard]] auto is_process_running(std::int64_t pid) -> bool
 {
-    ::HANDLE process = ::OpenProcess(PROCESS_QUERY_INFORMATION | PROCESS_VM_READ, FALSE, static_cast<DWORD>(pid));
+    ::HANDLE process = ::OpenProcess(
+        static_cast<DWORD>(PROCESS_QUERY_INFORMATION) | static_cast<DWORD>(PROCESS_VM_READ),
+        FALSE,
+        static_cast<DWORD>(pid));
 
     if (process != nullptr)
     {
